feat: rename Runtime classes, remove duplicated code, move to proper (sub)packages (wip)
