+++ plugin/src/main/scala-2/io/github/polentino/redacted/RedactedPlugin.scala
@@ -1,6 +1,8 @@
 package io.github.polentino.redacted
 
-import scala.tools.nsc._
+import io.github.polentino.redacted.components.RedactedPluginComponent
+
+import scala.tools.nsc.Global
 import scala.tools.nsc.plugins.{Plugin, PluginComponent}
 
 final class RedactedPlugin(override val global: Global) extends Plugin {
