--- plugin/src/main/scala/io/github/polentino/redacted/RedactedApi.scala
@@ -1,16 +1,0 @@
-package io.github.polentino.redacted
-
-trait RedactedApi[Api <: RuntimeApi] {
-  protected val runtime: Api
-
-  def process(tree: runtime.Tree): Option[runtime.Tree]
-}
-
-object RedactedApi {
-
-  def apply(api: RuntimeApi): RedactedApi[api.type] = new RedactedApi[api.type] {
-    override protected val runtime: api.type = api
-
-    override def process(tree: api.Tree): Option[api.Tree] = api.process(tree)
-  }
-}
