+++ plugin/src/main/scala/io/github/polentino/redacted/api/internal/ReporterApi.scala
@@ -1,4 +1,4 @@
-package io.github.polentino.redacted
+package io.github.polentino.redacted.api.internal
 
 trait ReporterApi[Api <: RuntimeApi] {
   protected val runtime: Api
